# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API cache/proxy — fetches from upstream REST APIs, stores in MongoDB, serves cached data to internal consumers.",
        "title": "cash-drugs API",
        "contact": {},
        "version": "0.10.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/cache/{slug}": {
            "get": {
                "description": "Returns cached upstream API data. Fetches from upstream if not cached.\nThe response data shape depends on the upstream API. For example, the\n`drugclasses` endpoint returns objects with fields: `name` (string),\n`type` (string), `code` (string), `codingSystem` (string).\nAvailable query parameters vary by endpoint — use `GET /api/endpoints`\nto discover each endpoint's supported params (e.g. BRAND_NAME, GENERIC_NAME,\nNDC, PHARM_CLASS for fda-ndc; SETID for spl-detail).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cache"
                ],
                "summary": "Get cached data for an endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Endpoint slug from config",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Path parameter value (varies by endpoint)",
                        "name": "SETID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Force upstream refresh (true/false)",
                        "name": "_force",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_cash-drugs_internal_model.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_cash-drugs_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_cash-drugs_internal_model.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service overloaded or upstream circuit open",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_cash-drugs_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_finish06_cash-drugs_internal_model.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "meta": {
                    "$ref": "#/definitions/github_com_finish06_cash-drugs_internal_model.ResponseMeta"
                }
            }
        },
        "github_com_finish06_cash-drugs_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "params": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "request_id": {
                    "type": "string"
                },
                "retry_after": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "upstream_status": {
                    "type": "integer"
                }
            }
        },
        "github_com_finish06_cash-drugs_internal_model.ResponseMeta": {
            "type": "object",
            "properties": {
                "fetched_at": {
                    "type": "string"
                },
                "page_count": {
                    "type": "integer"
                },
                "results_count": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "stale": {
                    "type": "boolean"
                },
                "stale_reason": {
                    "type": "string"
                }
            }
        },
...
    }
}